+++ R/listdial.R
@@ -73,17 +73,13 @@ dial <- function(socket,
                  url = "inproc://nanonext",
                  autostart = TRUE) {
 
-  if (missing(autostart) || isTRUE(autostart)) {
-    res <- .Call(rnng_dial, socket, url)
-    is.integer(res) && return(invisible(res))
-    if (.logging.) loginfo(evt = "dial start", pkey = "sock", pval = attr(res, "socket"),
-                           skey = "url", sval = url)
-
+  res <- if (missing(autostart) || isTRUE(autostart)) {
+    .Call(rnng_dial, socket, url)
   } else {
-    res <- .Call(rnng_dialer_create, socket, url)
-    is.integer(res) && return(invisible(res))
-
+    .Call(rnng_dialer_create, socket, url)
   }
+
+  is.integer(res) && return(invisible(res))
 
   if (is.environment(socket)) {
 
@@ -181,17 +177,13 @@ listen <- function(socket,
                    url = "inproc://nanonext",
                    autostart = TRUE) {
 
-  if (missing(autostart) || isTRUE(autostart)) {
-    res <- .Call(rnng_listen, socket, url)
-    is.integer(res) && return(invisible(res))
-    if (.logging.) loginfo(evt = "list start", pkey = "sock", pval = attr(res, "socket"),
-                           skey = "url", sval = url)
-
+  res <- if (missing(autostart) || isTRUE(autostart)) {
+    .Call(rnng_listen, socket, url)
   } else {
-    res <- .Call(rnng_listener_create, socket, url)
-    is.integer(res) && return(invisible(res))
-
+    .Call(rnng_listener_create, socket, url)
   }
+
+  is.integer(res) && return(invisible(res))
 
   if (is.environment(socket)) {