+++ R/socket.R
@@ -67,10 +67,6 @@ socket <- function(protocol = c("pair", "bus", "req", "rep", "push", "pull",
                                      "pub", "sub", "surveyor", "respondent"))
   res <- .Call(rnng_protocol_open, protocol)
   is.integer(res) && return(res)
-  if (.logging.) {
-    loginfo(evt = "sock open", pkey = "id", pval = attr(res, "id"),
-            skey = "protocol", sval = attr(res, "protocol"))
-  }
   if (!missing(dial)) dial(res, url = dial, autostart = autostart)
   if (!missing(listen)) listen(res, url = listen, autostart = autostart)
   res